Agregando mod double

+++ src/main/java/com/kodea/tdd/calc/oper/IOperador.java
@@ -2,11 +2,11 @@ package com.kodea.tdd.calc.oper;
 
 public interface IOperador {
 
-	double add(int val1, int val2);
+	double add(double val1, double val2);
 	
-	double subtract(int val1, int val2);
+	double subtract(double val1, double val2);
 	
-	double multiply(int val1, int val2);
+	double multiply(double val1, double val2);
 	
-	double divide(int val1, int val2);
+	double divide(double val1, double val2);
 }
